feat: move code around
